requirements.txt
python-dotenv==1.1.0
torch==2.6.0
fastapi==0.115.7
llama-index-core==0.12.28
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.3.30
nest-asyncio==1.6.0
openai==1.66.3
markitdown==0.1.1
python-multipart==0.0.18
uvicorn==0.30.6
markitdown[all]